[quickfix] fix frequency change crash
 Reason: when frame is created in createFrame(), the osc tables are set to point at the wavetables vector.
Unfortunately, when the new frame is returned from the function it is copy assigned or moved to a new frame at a new location.
Therefore, the address of the vector of wavetables changes and the osc has a dangling pointer which could point to anything.

Dirty solution: reinitialize wavetable after the frame has been push-backed to vector.

Warning: THIS WILL FUCK UP when the vector capacity is exceeded and all items are relocated!!!!!!!